+++ src/Test/Scene/wasmPhysicsTestScene.ts
@@ -234,6 +234,10 @@ export class SceneBuilder implements ISceneBuilder {
 
         for (const depthRenderer of Object.values(scene._depthRenderer)) {
             depthRenderer.forceDepthWriteTransparentMeshes = true;
+            engine.onResizeObservable.add(() => depthRenderer.getDepthMap().resize({
+                width: engine.getRenderWidth(),
+                height: engine.getRenderHeight()
+            }));
         }
 
         const video = document.createElement("video");
@@ -266,6 +270,10 @@ export class SceneBuilder implements ISceneBuilder {
                 });
             }
         });
+
+        // setTimeout(() => {
+        //     engine.snapshotRendering = true;
+        // }, 0);
 
         return scene;
     }